# openapi/openapi.yaml
openapi: 3.0.0
info:
    title: Midgard Public API
    version: 2.32.5
    contact:
        email: devs@thorchain.org
    description: >-
        The Midgard Public API queries THORChain and any chains linked via the Bifröst
        and prepares information about the network to be readily available for public users.
        The API parses transaction event data from THORChain and stores them
        in a time-series database to make time-dependent queries easy.
        Midgard does not hold critical information.
        To interact with THORChain protocol, users should query THORNode directly.
paths:
...
    "/v2/bonds/{address}":
        get:
            operationId: GetBondersDetails
            summary: Bonder Details
            description: |
                Returns the bond details of a given address.
                The response contains the bond amount, the bond status, and the bond type.
                This is wrapper from THORNode nodes API.
            parameters:
                - in: path
                  name: address
                  description: Address of the bonder
                  required: true
                  schema:
                      type: string
                  example: "thor102y0m3uptg0vvudeyh00r2fnz70wq7d8y7mu2g"
            responses:
                "200":
                    $ref: "#/components/responses/BonderDetailsResponse"
...
    "/v2/votes":
        get:
            operationId: GetVotes
            summary: Current Protocol Voting
            description: Returns timestamp, key, value for each active node member.
            parameters:
                - name: period
                  in: query
                  description: |
                      Specifies the base interval from which votes will be shown.
                      Default is 90d.
                  required: false
                  example: "90d"
                  schema:
                      type: string
                      enum:
                          [
                              "1h",
                              "24h",
                              "7d",
                              "14d",
                              "30d",
                              "90d",
                              "100d",
                              "180d",
                              "365d",
                              "all",
                          ]
            responses:
                "200":
                    $ref: "#/components/responses/VotesResponse"
...
    "/v2/pools":
        get:
            operationId: GetPools
            summary: Pools List
            description: Returns an array containing details for a set of pools
            parameters:
                - name: status
                  in: query
                  description: Filter for only pools with this status
                  required: false
                  schema:
                      type: string
                      enum: ["available", "staged", "suspended"]
                - name: period
                  in: query
                  description: |
                      Specifies the base interval from which annualPercentageRate and poolAPY is extrapolated.
                      Default is 14d.
                  required: false
                  example: "100d"
                  schema:
                      type: string
                      enum:
                          [
                              "1h",
                              "24h",
                              "7d",
                              "14d",
                              "30d",
                              "90d",
                              "100d",
                              "180d",
                              "365d",
                              "all",
                          ]
            responses:
                "200":
                    $ref: "#/components/responses/PoolsResponse"
    "/v2/pool/{asset}":
        get:
            operationId: GetPool
            summary: Details of a Pool
            description: "Returns details of the pool: depths, price, 24h volume, APY. "
            parameters:
                - in: path
                  name: asset
                  description: pool name
                  required: true
                  schema:
                      type: string
                  example: "BNB.TOMOB-1E1"
                - name: period
                  in: query
                  description: |
                      Specifies the base interval from which APY is extrapolated.
                      Default is 14d.
                  required: false
                  example: "100d"
                  schema:
                      type: string
                      enum:
                          [
                              "1h",
                              "24h",
                              "7d",
                              "14d",
                              "30d",
                              "90d",
                              "100d",
                              "180d",
                              "365d",
                              "all",
                          ]
            responses:
                "200":
                    $ref: "#/components/responses/PoolResponse"
    "/v2/pool/{asset}/stats":
        get:
            operationId: GetPoolStats
            summary: Pool Statistics
            description: |
                Statistics about the pool. The description of the fields have pointers about the
                corresponding v2/history location. Visit the history endpoint for drilldowns.
            parameters:
                - in: path
                  name: asset
                  description: pool name
                  required: true
                  schema:
                      type: string
                  example: "BNB.TOMOB-1E1"
                - name: period
                  in: query
                  description: |
                      Restricts aggregation type fields to the last period only.
                      Default is 14d.
                  required: false
                  example: "24h"
                  schema:
                      type: string
                      enum:
                          [
                              "1h",
                              "24h",
                              "7d",
                              "14d",
                              "30d",
                              "90d",
                              "100d",
                              "180d",
                              "365d",
                              "all",
                          ]
            responses:
                "200":
                    $ref: "#/components/responses/PoolStatsResponse"
    "/v2/history/depths/{pool}":
        get:
            operationId: GetDepthHistory
            summary: Depth and Price History
            description: |
                Returns the asset and rune depths and price.
                The values report the state at the end of each interval.

                History endpoint has two modes:
                * With Interval parameter it returns a series of time buckets. From and To dates will
                  be rounded to the Interval boundaries.
                * Without Interval parameter a single From..To search is performed with exact timestamps.


                * Interval: possible values: 5min, hour, day, week, month, quarter, year.
                * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
                * from/to: optional int, unix second.

                Possible usages with interval.
                * last 10 days: `?interval=day&count=10`
                * last 10 days before to: `?interval=day&count=10&to=1608825600`
                * next 10 days after from: `?interval=day&count=10&from=1606780800`
                * Days between from and to. From defaults to start of chain, to defaults to now.
                  Only the first 400 intervals are returned:
                  `interval=day&from=1606780800&to=1608825600`

                Pagination is possible with from&count and then using the returned meta.endTime as the
                From parameter of the next query.

                Possible configurations without interval:
                * exact search for one time frame: `?from=1606780899&to=1608825600`
                * one time frame until now: `?from=1606780899`
                * from chain start until now: no query parameters
            parameters:
                - name: pool
                  in: path
                  description: Return stats for this single pool.
                  required: true
                  schema:
                      type: string
                - name: interval
                  in: query
                  description: Interval of calculations
                  required: false
                  example: "day"
                  schema:
                      type: string
                      enum:
                          [
                              "5min",
                              "hour",
                              "day",
                              "week",
                              "month",
                              "quarter",
                              "year",
                          ]
                - name: count
                  in: query
                  description: Number of intervals to return. Should be between [1..400].
                  required: false
                  example: 30
                  schema:
                      type: integer
                - name: to
                  in: query
                  description: |
                      End time of the query as unix timestamp. If only count is given, defaults to now.
                  required: false
                  example: 1608825600
                  schema:
                      type: integer
                      format: int64
                - name: from
                  in: query
                  description: Start time of the query as unix timestamp
                  required: false
                  example: 1606780800
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    $ref: "#/components/responses/DepthHistoryResponse"
    "/v2/history/savers/{pool}":
        get:
            operationId: GetSaversHistory
            summary: Savers Units and Depth History
            description: |
                Returns savers depths and units.
                The values report the state at the end of each interval.

                History endpoint has two modes:
                * With Interval parameter it returns a series of time buckets. From and To dates will
                  be rounded to the Interval boundaries.
                * Without Interval parameter a single From..To search is performed with exact timestamps.


                * Interval: possible values: 5min, hour, day, week, month, quarter, year.
                * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
                * from/to: optional int, unix second.

                Possible usages with interval.
                * last 10 days: `?interval=day&count=10`
                * last 10 days before to: `?interval=day&count=10&to=1608825600`
                * next 10 days after from: `?interval=day&count=10&from=1606780800`
                * Days between from and to. From defaults to start of chain, to defaults to now.
                  Only the first 400 intervals are returned:
                  `interval=day&from=1606780800&to=1608825600`

                Pagination is possible with from&count and then using the returned meta.endTime as the
                From parameter of the next query.

                Possible configurations without interval:
                * exact search for one time frame: `?from=1606780899&to=1608825600`
                * one time frame until now: `?from=1606780899`
                * from chain start until now: no query parameters
            parameters:
                - name: pool
                  in: path
                  description: Return stats for this single pool.
                  required: true
                  schema:
                      type: string
                - name: interval
                  in: query
                  description: Interval of calculations
                  required: false
                  example: "day"
                  schema:
                      type: string
                      enum:
                          [
                              "5min",
                              "hour",
                              "day",
                              "week",
                              "month",
                              "quarter",
                              "year",
                          ]
                - name: count
                  in: query
                  description: Number of intervals to return. Should be between [1..400].
                  required: false
                  example: 30
                  schema:
                      type: integer
                - name: to
                  in: query
                  description: |
                      End time of the query as unix timestamp. If only count is given, defaults to now.
                  required: false
                  example: 1608825600
                  schema:
                      type: integer
                      format: int64
                - name: from
                  in: query
                  description: Start time of the query as unix timestamp
                  required: false
                  example: 1606780800
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    $ref: "#/components/responses/SaversHistoryResponse"
    "/v2/history/reserve":
        get:
            operationId: GetReserveHistory
            summary: Reserve income and expenses over bucketed history
            description: |
                Returns reserve module network fee, outbound fee, and gas reimbursement flow in
                bucketed history
            parameters:
                - name: interval
                  in: query
                  description: Interval of calculations
                  required: false
                  example: "day"
                  schema:
                      type: string
                      enum:
                          [
                              "5min",
                              "hour",
                              "day",
                              "week",
                              "month",
                              "quarter",
                              "year",
                          ]
                - name: count
                  in: query
                  description: Number of intervals to return. Should be between [1..400].
                  required: false
                  example: 30
                  schema:
                      type: integer
                - name: to
                  in: query
                  description: |
                      End time of the query as unix timestamp. If only count is given, defaults to now.
                  required: false
                  example: 1608825600
                  schema:
                      type: integer
                      format: int64
                - name: from
                  in: query
                  description: Start time of the query as unix timestamp
                  required: false
                  example: 1606780800
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    $ref: "#/components/responses/ReserveHistoryResponse"
    "/v2/history/rune":
        get:
            operationId: GetRunePriceHistory
            summary: Rune price history in USD
            description: |
                Returns rune price history in USD based on the time bucket given
            parameters:
                - name: interval
                  in: query
                  description: Interval of calculations
                  required: false
                  example: "day"
                  schema:
                      type: string
                      enum:
                          [
                              "5min",
                              "hour",
                              "day",
                              "week",
                              "month",
                              "quarter",
                              "year",
                          ]
                - name: count
                  in: query
                  description: Number of intervals to return. Should be between [1..400].
                  required: false
                  example: 30
                  schema:
                      type: integer
                - name: to
                  in: query
                  description: |
                      End time of the query as unix timestamp. If only count is given, defaults to now.
                  required: false
                  example: 1608825600
                  schema:
                      type: integer
                      format: int64
                - name: from
                  in: query
                  description: Start time of the query as unix timestamp
                  required: false
                  example: 1606780800
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    $ref: "#/components/responses/RunePriceHistoryResponse"
    "/v2/history/runepool":
        get:
            operationId: GetRUNEPoolHistory
            summary: RUNEPool total members and units History
            description: |
                Returns RUNEPool members and units.
                The values report the state at the end of each interval.

                History endpoint has two modes:
                * With Interval parameter it returns a series of time buckets. From and To dates will
                  be rounded to the Interval boundaries.
                * Without Interval parameter a single From..To search is performed with exact timestamps.


                * Interval: possible values: 5min, hour, day, week, month, quarter, year.
                * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
                * from/to: optional int, unix second.

                Possible usages with interval.
                * last 10 days: `?interval=day&count=10`
                * last 10 days before to: `?interval=day&count=10&to=1608825600`
                * next 10 days after from: `?interval=day&count=10&from=1606780800`
                * Days between from and to. From defaults to start of chain, to defaults to now.
                  Only the first 400 intervals are returned:
                  `interval=day&from=1606780800&to=1608825600`

                Pagination is possible with from&count and then using the returned meta.endTime as the
                From parameter of the next query.

                Possible configurations without interval:
                * exact search for one time frame: `?from=1606780899&to=1608825600`
                * one time frame until now: `?from=1606780899`
                * from chain start until now: no query parameters
            parameters:
                - name: interval
                  in: query
                  description: Interval of calculations
                  required: false
                  example: "day"
                  schema:
                      type: string
                      enum:
                          [
                              "5min",
                              "hour",
                              "day",
                              "week",
                              "month",
                              "quarter",
                              "year",
                          ]
                - name: count
                  in: query
                  description: Number of intervals to return. Should be between [1..400].
                  required: false
                  example: 30
                  schema:
                      type: integer
                - name: to
                  in: query
                  description: |
                      End time of the query as unix timestamp. If only count is given, defaults to now.
                  required: false
                  example: 1608825600
                  schema:
                      type: integer
                      format: int64
                - name: from
                  in: query
                  description: Start time of the query as unix timestamp
                  required: false
                  example: 1606780800
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    $ref: "#/components/responses/RUNEPoolHistoryResponse"
    "/v2/history/earnings":
        get:
            operationId: GetEarningsHistory
            summary: Earnings History
            description: |
                Returns earnings data for the specified interval.

                History endpoint has two modes:
                * With Interval parameter it returns a series of time buckets. From and To dates will
                  be rounded to the Interval boundaries.
                * Without Interval parameter a single From..To search is performed with exact timestamps.


                * Interval: possible values: 5min, hour, day, week, month, quarter, year.
                * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
                * from/to: optional int, unix second.

                Possible usages with interval.
                * last 10 days: `?interval=day&count=10`
                * last 10 days before to: `?interval=day&count=10&to=1608825600`
                * next 10 days after from: `?interval=day&count=10&from=1606780800`
                * Days between from and to. From defaults to start of chain, to defaults to now.
                  Only the first 400 intervals are returned:
                  `interval=day&from=1606780800&to=1608825600`

                Pagination is possible with from&count and then using the returned meta.endTime as the
                From parameter of the next query.

                Possible configurations without interval:
                * exact search for one time frame: `?from=1606780899&to=1608825600`
                * one time frame until now: `?from=1606780899`
                * from chain start until now: no query parameters
            parameters:
                - name: interval
                  in: query
                  description: Interval of calculations
                  required: false
                  example: "day"
                  schema:
                      type: string
                      enum:
                          [
                              "5min",
                              "hour",
                              "day",
                              "week",
                              "month",
                              "quarter",
                              "year",
                          ]
                - name: count
                  in: query
                  description: Number of intervals to return. Should be between [1..400].
                  required: false
                  example: 30
                  schema:
                      type: integer
                - name: to
                  in: query
                  description: |
                      End time of the query as unix timestamp. If only count is given, defaults to now.
                  required: false
                  example: 1608825600
                  schema:
                      type: integer
                      format: int64
                - name: from
                  in: query
                  description: Start time of the query as unix timestamp
                  required: false
                  example: 1606780800
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    $ref: "#/components/responses/EarningsHistoryResponse"
...
    "/v2/members":
        get:
            operationId: GetMembersAdresses
            summary: Members List
            description: |
                Returns an array containing the addresses for all pool members.
                Addresses are only shown once. If there's both a RUNE address and an asset address
                for a member, only the RUNE address will be shown.
            parameters:
                - name: pool
                  in: query
                  description: Return only members present in the pool.
                  required: false
                  schema:
                      type: string
            responses:
                "200":
                    "$ref": "#/components/responses/MembersResponse"
    "/v2/member/{address}":
        get:
            operationId: GetMemberDetail
            summary: Member Details
            description: |
                Returns an array of statistics for all the liquidity providers associated with a given
                member address.
            parameters:
                - name: address
                  in: path
                  description: |
                      Address to match liquidity providers. Either a rune or an asset address may be given.
                      Query can also be multiple addresses should be separated by comma (',')
                  required: true
                  schema:
                      type: string
                  example: "bnb1jxfh2g85q3v0tdq56fnevx6xcxtcnhtsmcu64m,bc1qcxssye4j6730h7ehgega3gyykkuwgdgmmpu62n"
                - name: showSavers
                  in: query
                  description: |
                      A flag to show saver vault membership details, the default is false.
                  required: false
                  example: true
                  schema:
                      type: boolean
                      default: false
            responses:
                "200":
                    $ref: "#/components/responses/MemberDetailsResponse"
    "/v2/borrowers":
        get:
            operationId: GetBorrowersAddresses
            summary: Borrowers List
            description: |
                Returns an array containing the addresses for all borrowers.
                Addresses are only shown once.
            parameters:
                - name: asset
                  in: query
                  description: Return only borrowers getting loan against this asset as collateral.
                  required: false
                  schema:
                      type: string
            responses:
                "200":
                    "$ref": "#/components/responses/BorrowersResponse"
    "/v2/borrower/{address}":
        get:
            operationId: GetBorrowerDetail
            summary: Borrower Details
            description: |
                Returns an array of statistics for all the open loans associated with a given
                borrower address.
            parameters:
                - name: address
                  in: path
                  description: |
                      Address to match borrower, an asset address is given.
                      Query can also be multiple addresses should be separated by comma (',')
                  required: true
                  schema:
                      type: string
                  example: "bnb1jxfh2g85q3v0tdq56fnevx6xcxtcnhtsmcu64m,bc1qcxssye4j6730h7ehgega3gyykkuwgdgmmpu62n"
            responses:
                "200":
                    $ref: "#/components/responses/BorrowerDetailsResponse"
...
components:
    responses:
...
        PoolsResponse:
            description: Array of pool details
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/PoolDetails"
...
        PoolStatsResponse:
            description: Stats for one pool.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/PoolStatsDetail"
        PoolResponse:
            description: Object containing details for one pool
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/PoolDetail"
        DepthHistoryResponse:
            description: Depth and price history
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/DepthHistory"
...
        BonderDetailsResponse:
            description: object containing bonder data for a specific member
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/BonderDetails"
...
        VotesResponse:
            description: Array containing the current protocol voting situation
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Votes"
...